Add product subscription

+++ HackDays.GraphQL/GraphQL/ProductMutation.cs
@@ -1,5 +1,6 @@
 ﻿using GraphQL;
 using GraphQL.Types;
+using HackDays.GraphQL.GraphQL.Messaging;
 using HackDays.GraphQL.GraphQL.Types;
 using HackDays.GraphQL.Models;
 using HackDays.GraphQL.Repositories;
@@ -8,15 +9,18 @@ namespace HackDays.GraphQL.GraphQL
 {
     public class ProductMutation : ObjectGraphType
     {
-        public ProductMutation(ProductRepository productRepository)
+        public ProductMutation(ProductRepository productRepository, MessagingService messagingService)
         {
             Field<ProductGraphType>(
                 "createProduct",
                 arguments: new QueryArguments(new QueryArgument<NonNullGraphType<ProductInputType>> { Name = "product" }),
                 resolve: context =>
                 {
-                    var product = context.GetArgument<Product>("product");
-                    return productRepository.Add(product);
+                    var product = productRepository.Add(context.GetArgument<Product>("product"));
+
+                    messagingService.AddProductMutatedMessage(product, "created");
+
+                    return product;
                 }
             );
 
@@ -30,11 +34,15 @@ namespace HackDays.GraphQL.GraphQL
                     var dbProduct = productRepository.GetById(product.Id);
                     if (dbProduct == null)
                     {
-                        context.Errors.Add(new ExecutionError("Couldn't find product in db."));
+                        context.Errors.Add(new ExecutionError("Couldn't find product in database."));
                         return null;
                     }
 
-                    return productRepository.Update(product);
+                    var updatedProduct = productRepository.Update(product);
+
+                    messagingService.AddProductMutatedMessage(product, "updated");
+
+                    return updatedProduct;
                 }
             );
 
@@ -45,7 +53,18 @@ namespace HackDays.GraphQL.GraphQL
                 ),
                 resolve: context => {
                     var id = context.GetArgument<int>("id");
-                    return productRepository.Delete(id);
+                    var product = productRepository.GetById(id);
+                    if (product == null)
+                    {
+                        context.Errors.Add(new ExecutionError("Couldn't find product in db."));
+                        return false;
+                    }
+
+                    productRepository.Delete(id);
+
+                    messagingService.AddProductMutatedMessage(product, "deleted");
+
+                    return true;
                 }
             );
         }
